microservices/services/resources-fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.5.1
